requirements.txt
astroid==2.2.5
cffi==1.12.3
Django==2.2.5
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.2
Pillow==6.1.0
pycparser==2.19
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
webp==0.1.0a13
wrapt==1.11.2
